feat: add error handling

+++ app/main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI,UploadFile
+from fastapi import FastAPI, HTTPException,UploadFile
 from app.services.ocr_service import extract_text_from_pdf
 from app.services.llm_service import extract_data
 import os
@@ -13,13 +13,28 @@ def read_health():
 
 @app.post("/extract")
 def extract(file: UploadFile):
+    if file.content_type != "application/pdf":
+        raise HTTPException(status_code=400, detail={"error": "Invalid file type. Please upload a PDF."})
+
     os.makedirs("temp", exist_ok=True)
     file_path=f"temp/{file.filename}"
     with open(file_path, "wb") as f:
         f.write(file.file.read())
     ocrText = extract_text_from_pdf(file_path) # extract text from pdf and return the result
-    print("text: ",ocrText)
-    data = extract_data(ocrText) # extract data from text and return the result
-    os.remove(file_path)    
+
+    if(ocrText is None or ocrText.strip() == ""):
+        os.remove(file_path)
+        raise HTTPException(status_code=400, detail={"error": "Unable to extract text from the PDF. Please ensure the document is clear and try again."})
+    
+    try:
+
+        data = extract_data(ocrText) # extract data from text and return the result
+
+    except ValueError as e:
+        os.remove(file_path)
+        raise HTTPException(status_code=422, detail={"error": str(e)})
+
+
+    os.remove(file_path)
     return data
